refactor : Piece 기능 추가
